+++ NEWS.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## Unreleased
+## [1.4.0] - 2018-07-03
 ### Added
 - `split` argument for separating columns (useful for interactions)
 - how to plot interactions in vignette
@@ -63,6 +63,7 @@
 - `create_start_event` function
 - `plot_gamsd` function
 
+[1.4.0]: https://github.com/stefanocoretta/tidymv/compare/v1.3.1...v1.4.0
 [1.3.1]: https://github.com/stefanocoretta/tidymv/compare/v1.3.0...v1.3.1
 [1.3.0]: https://github.com/stefanocoretta/tidymv/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/stefanocoretta/tidymv/compare/v1.1.0...v1.2.0